+++ pyserini/encode/_pca.py
@@ -1,0 +1,34 @@
+#
+# Pyserini: Reproducible IR research with sparse and dense representations
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import faiss
+import numpy as np
+
+
+class PcaEncoder:
+    def __init__(self, encoder, pca_model_path):
+        self.encoder = encoder
+        self.pca_mat = faiss.read_VectorTransform(pca_model_path)
+
+    def encode(self, text, **kwargs):
+        if isinstance(text, str):
+            embeddings = self.encoder.encode(text, **kwargs)
+            embeddings = self.pca_mat.apply_py(np.array([embeddings]))
+            embeddings = embeddings[0]
+        else:
+            embeddings = self.encoder.encode(text, **kwargs)
+            embeddings = self.pca_mat.apply_py(embeddings)
+        return embeddings
